move secrets management to 'system-tests'
