+++ core/scripts/cre/environment/environment/workflow.go
@@ -2,23 +2,16 @@ package environment
 
 import (
 	"context"
-	"encoding/hex"
-	"encoding/json"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
 	"slices"
-	"strconv"
 	"time"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
-	"gopkg.in/yaml.v2"
 
-	secretsUtils "github.com/smartcontractkit/chainlink-common/pkg/workflows/secrets"
-	"github.com/smartcontractkit/chainlink-evm/gethwrappers/keystone/generated/capabilities_registry"
 	"github.com/smartcontractkit/chainlink-testing-framework/framework/components/blockchain"
 	"github.com/smartcontractkit/chainlink-testing-framework/seth"
 
@@ -298,55 +291,17 @@ func compileCopyAndRegisterWorkflow(ctx context.Context, workflowFilePathFlag, w
 
 	var secretsPath *string
 	if secretsFilePathFlag != "" {
-		fmt.Printf("\n⚙️ Loading workflow secrets\n")
+		fmt.Printf("\n⚙️ Loading and encrypting workflow secrets\n")
 
-		secretsConfig, err := newSecretsConfig(secretsFilePathFlag)
-		if err != nil {
-			return err
+		secretPathAbs, secretsErr := creworkflow.PrepareSecrets(sethClient, uint32(donIDFlag), common.HexToAddress(capabilitiesRegistryAddressFlag), common.HexToAddress(workflowOwnerAddressFlag), secretsFilePathFlag)
+		if secretsErr != nil {
+			return errors.Wrap(secretsErr, "failed to prepare secrets")
 		}
 
-		envSecrets, err := loadSecretsFromEnvironment(secretsConfig)
-		if err != nil {
-			return err
-		}
-
-		fmt.Printf("\n✅ Loaded workflow secrets\n\n")
-
-		fmt.Printf("\n⚙️ Encrypting workflow secrets\n")
-
-		encryptSecrets, err := encryptSecrets(sethClient, common.HexToAddress(capabilitiesRegistryAddressFlag), donIDFlag, workflowOwnerAddressFlag, envSecrets, secretsConfig)
-		if err != nil {
-			return err
-		}
-
-		fmt.Printf("\n✅ Encrypted workflow secrets\n\n")
-
-		fmt.Printf("\n⚙️ Writing encrypted secrets file to disk\n")
-
-		encryptedSecretsFilePath := "./encrypted.secrets.json"
-		encryptedSecretsFile, err := os.Create(encryptedSecretsFilePath)
-		if err != nil {
-			return fmt.Errorf("failed to create secrets file: %w", err)
-		}
-		defer encryptedSecretsFile.Close()
-		defer func() {
-			_ = os.Remove(encryptedSecretsFilePath)
-		}()
-
-		encoder := json.NewEncoder(encryptedSecretsFile)
-		if err := encoder.Encode(encryptSecrets); err != nil {
-			return fmt.Errorf("failed to write to secrets file: %w", err)
-		}
-
-		fmt.Printf("\n✅ Wrote encrypted secrets file to disk\n\n")
+		fmt.Printf("\n✅ Encrypted workflow secrets file prepared\n\n")
 
 		fmt.Printf("\n⚙️ Copying encrypted secrets file to Docker container\n")
-		secretPathAbs, secretPathAbsErr := filepath.Abs(encryptedSecretsFilePath)
-		if secretPathAbsErr != nil {
-			return errors.Wrap(secretPathAbsErr, "failed to get absolute path of the encrypted secrets file")
-		}
-
-		secretsCopyErr := creworkflow.CopyWorkflowToDockerContainers(encryptedSecretsFilePath, containerNamePatternFlag, containerTargetDirFlag)
+		secretsCopyErr := creworkflow.CopyWorkflowToDockerContainers(secretPathAbs, containerNamePatternFlag, containerTargetDirFlag)
 		if secretsCopyErr != nil {
 			return errors.Wrap(secretsCopyErr, "❌ failed to copy encrypted secrets file to Docker container")
 		}
@@ -354,7 +309,7 @@ func compileCopyAndRegisterWorkflow(ctx context.Context, workflowFilePathFlag, w
 		secretPathAbs = "file://" + secretPathAbs
 		secretsPath = &secretPathAbs
 
-		fmt.Printf("\n✅ Workflow encrypted secrets file copied to Docker container\n\n")
+		fmt.Printf("\n✅ Encrypted workflow secrets file copied to Docker container\n\n")
 	}
 
 	fmt.Printf("\n⚙️ Deleting workflow '%s' from the workflow registry\n\n", workflowNameFlag)
@@ -389,96 +344,4 @@ func compileCopyAndRegisterWorkflow(ctx context.Context, workflowFilePathFlag, w
 	fmt.Printf("\n✅ Workflow registered successfully\n\n")
 
 	return nil
-}
-
-func newSecretsConfig(configPath string) (*secretsUtils.SecretsConfig, error) {
-	secretsConfigFile, err := os.Open(configPath)
-	if err != nil {
-		return nil, fmt.Errorf("error opening secrets config file: %w", err)
-	}
-	defer secretsConfigFile.Close()
-
-	var config secretsUtils.SecretsConfig
-	err = yaml.NewDecoder(secretsConfigFile).Decode(&config)
-	if err != nil && errors.Is(err, io.EOF) {
-		return &secretsUtils.SecretsConfig{}, nil
-	}
-	if err != nil {
-		return nil, fmt.Errorf("error decoding secrets config file: %w", err)
-	}
-
-	return &config, nil
-}
-
-func loadSecretsFromEnvironment(config *secretsUtils.SecretsConfig) (map[string][]string, error) {
-	secrets := make(map[string][]string)
-	for secretName, envVars := range config.SecretsNames {
-		for _, envVar := range envVars {
-			secretValue := os.Getenv(envVar)
-			if secretValue == "" {
-				return nil, fmt.Errorf("missing environment variable: %s", envVar)
-			}
-			secrets[secretName] = append(secrets[secretName], secretValue)
-		}
-	}
-	return secrets, nil
-}
-
-func encryptSecrets(c *seth.Client, capabilitiesRegistry common.Address, donID uint32, workflowOwner string, secrets map[string][]string, config *secretsUtils.SecretsConfig) (secretsUtils.EncryptedSecretsResult, error) {
-	cr, err := capabilities_registry.NewCapabilitiesRegistry(capabilitiesRegistry, c.Client)
-	if err != nil {
-		return secretsUtils.EncryptedSecretsResult{}, fmt.Errorf("failed to attach to the Capabilities Registry contract: %w", err)
-	}
-
-	nodeInfos, err := cr.GetNodes(c.NewCallOpts())
-	if err != nil {
-		return secretsUtils.EncryptedSecretsResult{}, fmt.Errorf("failed to get node information from the Capabilities Registry contract: %w", err)
-	}
-
-	donInfo, err := cr.GetDON(c.NewCallOpts(), donID)
-	if err != nil {
-		return secretsUtils.EncryptedSecretsResult{}, fmt.Errorf("failed to get DON information from the Capabilities Registry contract: %w", err)
-	}
-
-	encryptionPublicKeys := make(map[string][32]byte)
-	for _, nodeInfo := range nodeInfos {
-		// Filter only the nodes that are part of the DON
-		if secretsUtils.ContainsP2pId(nodeInfo.P2pId, donInfo.NodeP2PIds) {
-			encryptionPublicKeys[hex.EncodeToString(nodeInfo.P2pId[:])] = nodeInfo.EncryptionPublicKey
-		}
-	}
-
-	if len(encryptionPublicKeys) == 0 {
-		return secretsUtils.EncryptedSecretsResult{}, errors.New("no nodes found for the don")
-	}
-
-	// Encrypt secrets for each node
-	encryptedSecrets, secretsEnvVarsByNode, err := secretsUtils.EncryptSecretsForNodes(
-		workflowOwner,
-		secrets,
-		encryptionPublicKeys,
-		secretsUtils.SecretsConfig{SecretsNames: config.SecretsNames},
-	)
-	if err != nil {
-		return secretsUtils.EncryptedSecretsResult{}, fmt.Errorf("node public keys not found: %w", err)
-	}
-
-	// Convert encryptionPublicKey to hex strings for including in the metadata
-	nodePublicEncryptionKeys := make(map[string]string)
-	for p2pID, encryptionPublicKey := range encryptionPublicKeys {
-		nodePublicEncryptionKeys[p2pID] = hex.EncodeToString(encryptionPublicKey[:])
-	}
-
-	result := secretsUtils.EncryptedSecretsResult{
-		EncryptedSecrets: encryptedSecrets,
-		Metadata: secretsUtils.Metadata{
-			WorkflowOwner:            workflowOwner,
-			CapabilitiesRegistry:     capabilitiesRegistry.String(),
-			DonId:                    strconv.FormatUint(uint64(donID), 10),
-			DateEncrypted:            time.Now().Format(time.RFC3339),
-			NodePublicEncryptionKeys: nodePublicEncryptionKeys,
-			EnvVarsAssignedToNodes:   secretsEnvVarsByNode,
-		},
-	}
-	return result, nil
 }
